Fixed review download error when topic is empty
I think I tracked down the cause of a problem. It is about an exception handler, which is incomplete. 

My problem occurred only if I had no title set in Gerrit and let to a situation in which `git review -d [CHANGE]` tried to
`git clone branch/user/`
as opposed to
`git clone branch/user`
That leads to an error because after a slash usually follows the topic and if that is empty the checkout does not work as expected. It instead stops with an error message `remote branch not found`.